This optimises the construction of `DynTiles`, which was is taking almost all the time (just filling memory with zeros) syncing Taurion with empty map.  Instead of always using a full `std::vector<bool>` for all map tiles, we use a vector of bitsets and only initialise those "buckets" that are actually updated from all-default values.

This reduces the construction time according to the benchmark massively from 2086us to 1.38us, while only increasing times of other operations moderately by 30-50% in the benchmark cases.  (And in any case, these times are irrelevant unless we have millions of players on the map and the bottlenecks will be elsewhere.)

With this, the typical full processing time per block while syncing (with a mostly empty map) for me reduces from 12ms to 150us.  The profile of syncing time afterwards looks a lot more balanced; still `DynObstacles` takes almost all time, but mostly in querying the database for vehicles rather than just filling memory with zeros.